Create Find Anagram Pairs in Sentence.java
You are given a sentence containing multiple words. Your task is to find and print all pairs of words that are anagrams of each other. • Each pair should be printed in the order the original word appears in the input. • Each anagram should only be part of one pair. • If no such pairs exist, print "No Anagrams Found".